+++ BleWinrtDll/logging.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void clearError();
+void saveError(const wchar_t* message, ...);
